+++ components/engine_panic/src/engine.rs
@@ -2,7 +2,7 @@
 
 use engine_traits::{
     IterMetricsCollector, IterOptions, Iterable, Iterator, KvEngine, MetricsExt, Peekable,
-    ReadOptions, Result, SnapshotContext, SyncMutable, WriteOptions,
+    ReadOptions, Result, SyncMutable, WriteOptions,
 };
 
 use crate::{db_vector::PanicDbVector, snapshot::PanicSnapshot, write_batch::PanicWriteBatch};
