+++ components/hybrid_engine/src/engine.rs
@@ -1,13 +1,6 @@
 // Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
 
-use engine_traits::{
-    KvEngine, Mutable, Peekable, ReadOptions, RegionCacheEngine, Result, SnapshotContext,
-    SnapshotMiscExt, SyncMutable, WriteBatch, WriteBatchExt,
-};
-use keys::DATA_PREFIX_KEY;
-use kvproto::metapb::{self, RegionEpoch};
-
-use crate::snapshot::HybridEngineSnapshot;
+use engine_traits::{CacheRegion, KvEngine, RegionCacheEngine};
 
 /// This engine is structured with both a disk engine and an region cache
 /// engine. The disk engine houses the complete database data, whereas the
@@ -30,28 +23,6 @@ where
     EK: KvEngine,
     EC: RegionCacheEngine,
 {
-    pub fn disk_engine(&self) -> &EK {
-        &self.disk_engine
-    }
-
-    pub fn mut_disk_engine(&mut self) -> &mut EK {
-        &mut self.disk_engine
-    }
-
-    pub fn region_cache_engine(&self) -> &EC {
-        &self.region_cache_engine
-    }
-
-    pub fn mut_region_cache_engine(&mut self) -> &mut EC {
-        &mut self.region_cache_engine
-    }
-}
-
-impl<EK, EC> HybridEngine<EK, EC>
-where
-    EK: KvEngine,
-    EC: RegionCacheEngine,
-{
     pub fn new(disk_engine: EK, region_cache_engine: EC) -> Self {
         Self {
             disk_engine,
@@ -59,7 +30,12 @@ where
         }
     }
 
-    pub fn new_snapshot(&self, ctx: Option<SnapshotContext>) -> HybridEngineSnapshot<EK, EC> {
+    #[cfg(test)]
+    pub(crate) fn new_snapshot(
+        &self,
+        ctx: Option<SnapshotContext>,
+    ) -> crate::HybridEngineSnapshot<EK, EC> {
+        use engine_traits::SnapshotMiscExt;
         let disk_snap = self.disk_engine.snapshot();
         let region_cache_snap = if !self.region_cache_engine.enabled() {
             None
@@ -74,127 +50,37 @@ where
         } else {
             None
         };
-        HybridEngineSnapshot::new(disk_snap, region_cache_snap)
+        crate::HybridEngineSnapshot::new(disk_snap, region_cache_snap)
+    }
+
+    pub(crate) fn disk_engine(&self) -> &EK {
+        &self.disk_engine
+    }
+
+    pub fn region_cache_engine(&self) -> &EC {
+        &self.region_cache_engine
     }
 }
 
-impl<EK, EC> HybridEngine<EK, EC>
-where
-    EK: KvEngine,
-    EC: RegionCacheEngine,
-    HybridEngine<EK, EC>: WriteBatchExt,
-{
-    fn sync_write<F>(&self, key: &[u8], f: F) -> Result<()>
-    where
-        F: FnOnce(&mut <Self as WriteBatchExt>::WriteBatch) -> Result<()>,
-    {
-        let mut batch = self.write_batch();
-        if let Some(cached_region) = self.region_cache_engine.get_region_for_key(key) {
-            // CacheRegion does not contains enough information for Region so the transfer
-            // is not accurate but this method is not called in production code.
-            let mut region = metapb::Region::default();
-            region.set_id(cached_region.id);
-            region.set_start_key(cached_region.start[DATA_PREFIX_KEY.len()..].to_vec());
-            region.set_end_key(cached_region.end[DATA_PREFIX_KEY.len()..].to_vec());
-            let mut epoch = RegionEpoch::default();
-            epoch.version = cached_region.epoch_version;
-            region.set_region_epoch(epoch);
-            batch.prepare_for_region(&region);
-        }
-        f(&mut batch)?;
-        let _ = batch.write()?;
-        Ok(())
-    }
+#[derive(Debug, Clone)]
+pub struct SnapshotContext {
+    pub region: Option<CacheRegion>,
+    pub read_ts: u64,
 }
 
-// todo: implement KvEngine methods as well as it's super traits.
-impl<EK, EC> KvEngine for HybridEngine<EK, EC>
-where
-    EK: KvEngine,
-    EC: RegionCacheEngine,
-    HybridEngine<EK, EC>: WriteBatchExt,
-{
-    type Snapshot = HybridEngineSnapshot<EK, EC>;
-
-    fn snapshot(&self) -> Self::Snapshot {
-        unreachable!()
-    }
-
-    fn sync(&self) -> engine_traits::Result<()> {
-        self.disk_engine.sync()
-    }
-
-    fn bad_downcast<T: 'static>(&self) -> &T {
-        self.disk_engine.bad_downcast()
-    }
-
-    #[cfg(feature = "testexport")]
-    fn inner_refcount(&self) -> usize {
-        self.disk_engine.inner_refcount()
-    }
-}
-
-impl<EK, EC> Peekable for HybridEngine<EK, EC>
-where
-    EK: KvEngine,
-    EC: RegionCacheEngine,
-{
-    type DbVector = EK::DbVector;
-
-    // region cache engine only supports peekable trait in the snapshot of it
-    fn get_value_opt(&self, opts: &ReadOptions, key: &[u8]) -> Result<Option<Self::DbVector>> {
-        self.disk_engine.get_value_opt(opts, key)
-    }
-
-    // region cache engine only supports peekable trait in the snapshot of it
-    fn get_value_cf_opt(
-        &self,
-        opts: &ReadOptions,
-        cf: &str,
-        key: &[u8],
-    ) -> Result<Option<Self::DbVector>> {
-        self.disk_engine.get_value_cf_opt(opts, cf, key)
-    }
-}
-
-impl<EK, EC> SyncMutable for HybridEngine<EK, EC>
-where
-    EK: KvEngine,
-    EC: RegionCacheEngine,
-    HybridEngine<EK, EC>: WriteBatchExt,
-{
-    fn put(&self, key: &[u8], value: &[u8]) -> Result<()> {
-        self.sync_write(key, |b| b.put(key, value))
-    }
-
-    fn put_cf(&self, cf: &str, key: &[u8], value: &[u8]) -> Result<()> {
-        self.sync_write(key, |b| b.put_cf(cf, key, value))
-    }
-
-    fn delete(&self, key: &[u8]) -> Result<()> {
-        self.sync_write(key, |b| b.delete(key))
-    }
-
-    fn delete_cf(&self, cf: &str, key: &[u8]) -> Result<()> {
-        self.sync_write(key, |b| b.delete_cf(cf, key))
-    }
-
-    fn delete_range(&self, begin_key: &[u8], end_key: &[u8]) -> Result<()> {
-        self.sync_write(begin_key, |b| b.delete_range(begin_key, end_key))
-    }
-
-    fn delete_range_cf(&self, cf: &str, begin_key: &[u8], end_key: &[u8]) -> Result<()> {
-        self.sync_write(begin_key, |b| b.delete_range_cf(cf, begin_key, end_key))
+impl SnapshotContext {
+    pub fn set_region(&mut self, region: CacheRegion) {
+        assert!(self.region.is_none());
+        self.region = Some(region);
     }
 }
 
 #[cfg(test)]
 mod tests {
-
     use std::sync::Arc;
 
     use engine_rocks::util::new_engine;
-    use engine_traits::{CacheRegion, SnapshotContext, CF_DEFAULT, CF_LOCK, CF_WRITE};
+    use engine_traits::{CacheRegion, CF_DEFAULT, CF_LOCK, CF_WRITE};
     use in_memory_engine::{
         config::InMemoryEngineConfigManager, test_util::new_region, InMemoryEngineConfig,
         InMemoryEngineContext, RegionCacheMemoryEngine,
@@ -203,6 +89,7 @@ mod tests {
     use tempfile::Builder;
     use tikv_util::config::VersionTrack;
 
+    use super::*;
     use crate::HybridEngine;
 
     #[test]
