+++ components/test_raftstore/src/cluster.rs
@@ -2059,27 +2059,3 @@ impl RawEngine<RocksEngine> for RocksEngine {
         self.get_msg_cf(CF_RAFT, &keys::raft_state_key(region_id))
     }
 }
-
-impl RawEngine<RocksEngine> for HybridEngineImpl {
-    fn region_cache_engine(&self) -> bool {
-        true
-    }
-
-    fn region_local_state(
-        &self,
-        region_id: u64,
-    ) -> engine_traits::Result<Option<RegionLocalState>> {
-        self.disk_engine()
-            .get_msg_cf(CF_RAFT, &keys::region_state_key(region_id))
-    }
-
-    fn raft_apply_state(&self, region_id: u64) -> engine_traits::Result<Option<RaftApplyState>> {
-        self.disk_engine()
-            .get_msg_cf(CF_RAFT, &keys::apply_state_key(region_id))
-    }
-
-    fn raft_local_state(&self, region_id: u64) -> engine_traits::Result<Option<RaftLocalState>> {
-        self.disk_engine()
-            .get_msg_cf(CF_RAFT, &keys::raft_state_key(region_id))
-    }
-}
